Renamed to vulnerablecomponents
