+++ src/main/java/org/owasp/webgoat/lessons/vulnerablecomponents/Contact.java
@@ -20,16 +20,16 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.vulnerable_components;
+package org.owasp.webgoat.lessons.vulnerablecomponents;
 
-import lombok.Data;
+public interface Contact {
 
-@Data
-public class ContactImpl implements Contact {
-
-	private Integer id;
-	private String firstName;
-	private String lastName;
-	private String email;
-
+	public Integer getId();
+	public void setId(Integer id);
+	public String getFirstName();
+	public void setFirstName(String firstName);
+	public String getLastName();
+	public void setLastName(String lastName);
+	public String getEmail();
+	public void setEmail(String email);
 }
