+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content0.adoc
