+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content1.adoc
