+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content2.adoc
