+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content3.adoc
