+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content4a.adoc
