+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content5.adoc
