+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_content6.adoc
