+++ src/main/resources/lessons/vulnerablecomponents/documentation/VulnerableComponents_plan.adoc
