+++ src/main/resources/lessons/vulnerablecomponents/i18n/WebGoatLabels.properties
