+++ src/main/resources/lessons/vulnerablecomponents/images/OWASP-2013-A9.png
