+++ src/main/resources/lessons/vulnerablecomponents/images/OWASP-Dep-Check.png
