+++ src/main/resources/lessons/vulnerablecomponents/images/Old-Components.png
