+++ src/main/resources/lessons/vulnerablecomponents/images/Risk-of-Old-Components.png
