+++ src/main/resources/lessons/vulnerablecomponents/images/WebGoat-Vulns.png
